refactor: moved handleMoveInstance and handleDeleteInstance to a separate file
